DegenPrimer/DegenPrimerConfig.py: corrected some help strings.
PCR_Simulation: changed the way histogram is represented. Each row is multiline now in order to display all the information about a product. Also to the information about a product the list of primers which gave it is added. Products are filtered by quantity now. Description of histogram was updated.

Primer: fixed bug in find_records method.

SearchEngine: _find_mp and plain _find are now private methods. New public methods -- find, which chooses whether to use mp or not, and batch_find -- are added.

StringTools.py: added line_by_line function which format given strings in a series of cells of given widths.